backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
openai==1.3.0
textblob==0.17.1
pandas==2.1.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
isodate==0.6.1
gunicorn==21.2.0